+++ android/src/main/java/io/github/zileyuan/flutter_walle/FlutterWallePlugin.java
@@ -1,31 +1,43 @@
 package io.github.zileyuan.flutter_walle;
 
+import androidx.annotation.NonNull;
+
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
+import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import com.meituan.android.walle.WalleChannelReader;
 import android.content.Context;
 import android.text.TextUtils;
 
 /** FlutterWallePlugin */
-public class FlutterWallePlugin implements MethodCallHandler {
+public class FlutterWallePlugin implements FlutterPlugin, MethodCallHandler {
 
-  private Context context;
+  private static Context context;
+  public static MethodChannel methodChannel;
 
-  private FlutterWallePlugin(Context ctx) {
-    context = ctx;
-  }
-  
-  /** Plugin registration. */
-  public static void registerWith(Registrar registrar) {
-    final MethodChannel channel = new MethodChannel(registrar.messenger(), "flutter_walle");
-    channel.setMethodCallHandler(new FlutterWallePlugin(registrar.context()));
+  @Override
+  public void onAttachedToEngine(@NonNull FlutterPluginBinding flutterPluginBinding) {
+    context = flutterPluginBinding.getApplicationContext();
+    methodChannel = new MethodChannel(flutterPluginBinding.getFlutterEngine().getDartExecutor(), "flutter_walle");
+    methodChannel.setMethodCallHandler(this);
   }
 
   @Override
-  public void onMethodCall(MethodCall call, Result result) {
+  public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
+    methodChannel.setMethodCallHandler(null);
+  }
+
+  public static void registerWith(Registrar registrar) {
+    context = registrar.context();
+    methodChannel = new MethodChannel(registrar.messenger(), "flutter_walle");
+    methodChannel.setMethodCallHandler(new FlutterWallePlugin());
+  }
+
+  @Override
+  public void onMethodCall(@NonNull MethodCall call, @NonNull Result result) {
     if (call.method.equals("getChannel")) {
       String channel = WalleChannelReader.getChannel(context);
       if (TextUtils.isEmpty(channel)) {
@@ -35,13 +47,5 @@ public class FlutterWallePlugin implements MethodCallHandler {
     } else {
       result.notImplemented();
     }
-  }
-
-  public static String getChannel(Context ctx) {
-    String channel = WalleChannelReader.getChannel(ctx);
-    if (channel == null || channel.length() == 0) {
-        channel = "NoChannel";
-    }
-    return channel;
   }
 }