+++ lib/flutter_walle.dart
@@ -3,9 +3,10 @@ import 'dart:async';
 import 'package:flutter/services.dart';
 
 class FlutterWalle {
-  static const MethodChannel _channel =
-      const MethodChannel('flutter_walle');
+  /// Communication MethodChannel
+  static const MethodChannel _channel = const MethodChannel('flutter_walle');
 
+  /// Get a channel from current package
   static Future<String> get channel async {
     final String ch = await _channel.invokeMethod('getChannel');
     return ch;