+++ CHANGELOG.md
@@ -1,7 +1,19 @@
+## 1.0.02
+
+- Support new plug-in system.
+- 支持新插件系统
+
+## 1.0.01
+
+- Add code comments, format code.
+- 加入代码注释，格式化代码
+
 ## 0.0.2
 
-* Add some description to find easy.
+- Add some description to find easy.
+- 加入一些描述更容易发现
 
 ## 0.0.1
 
-* Implements meituan walle read channel.
+- Implements meituan walle read channel.
+- 实现美团瓦力打包读取渠道功能